tests: Rune sign mint order op tests
