+++ src/rune-bridge/src/core/deposit.rs
@@ -4,7 +4,6 @@ use std::rc::Rc;
 
 use bitcoin::hashes::Hash;
 use bitcoin::{Address, Network};
-use bridge_canister::bridge::OperationContext;
 use bridge_canister::runtime::RuntimeState;
 use bridge_did::id256::Id256;
 use bridge_did::order::{MintOrder, SignedMintOrder};
@@ -373,21 +372,6 @@ impl<UTXO: UtxoProvider, INDEX: RuneIndexProvider> RuneDeposit<UTXO, INDEX> {
             approve_amount: Default::default(),
             fee_payer: H160::default(),
         }
-    }
-
-    pub async fn sign_mint_order(
-        &self,
-        mint_order: MintOrder,
-    ) -> Result<SignedMintOrder, DepositError> {
-        let signer = self.runtime_state.get_signer().map_err(|err| {
-            DepositError::Unavailable(format!("cannot initialize signer: {err:?}"))
-        })?;
-        let signed_mint_order = mint_order
-            .encode_and_sign(&signer)
-            .await
-            .map_err(|err| DepositError::Sign(format!("{err:?}")))?;
-
-        Ok(signed_mint_order)
     }
 
     fn filter_out_used_utxos(&self, get_utxos_response: &mut GetUtxosResponse) {
