+++ src/rune-bridge/src/ops.rs
@@ -10,6 +10,7 @@ use bridge_utils::bft_events::{
 };
 use candid::{CandidType, Decode, Deserialize};
 use did::{H160, H256};
+use eth_signer::sign_strategy::TransactionSigner;
 use ic_exports::ic_cdk::api::management_canister::bitcoin::Utxo;
 use ic_task_scheduler::task::TaskOptions;
 use serde::Serialize;
@@ -112,7 +113,8 @@ impl Operation for RuneBridgeOp {
                 mint_order,
             } => {
                 log::debug!("RuneBridgeOp::SignMintOrder {dst_address} {mint_order:?}");
-                Self::sign_mint_order(ctx, id.nonce(), dst_address, mint_order).await
+                let signer = ctx.get_signer()?;
+                Self::sign_mint_order(&signer, id.nonce(), dst_address, mint_order).await
             }
             RuneBridgeOp::SendMintOrder { dst_address, order } => {
                 log::debug!("RuneBridgeOp::SendMintOrder {dst_address} {order:?}");
@@ -417,7 +419,7 @@ impl RuneBridgeOp {
     }
 
     async fn sign_mint_order(
-        ctx: RuntimeState<Self>,
+        signer: &impl TransactionSigner,
         nonce: u32,
         dst_address: H160,
         mut mint_order: MintOrder,
@@ -425,11 +427,7 @@ impl RuneBridgeOp {
         // update nonce
         mint_order.nonce = nonce;
 
-        let deposit = RuneDeposit::get(ctx);
-        let signed = deposit
-            .sign_mint_order(mint_order)
-            .await
-            .map_err(|err| Error::FailedToProgress(format!("cannot sign mint order: {err:?}")))?;
+        let signed = mint_order.encode_and_sign(signer).await?;
 
         Ok(Self::SendMintOrder {
             dst_address,
@@ -490,555 +488,4 @@ pub struct RuneDepositRequestData {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::cell::RefCell;
-    use std::rc::Rc;
-    use std::str::FromStr;
-
-    use bridge_canister::memory::{memory_by_id, StableMemory};
-    use bridge_canister::operation_store::OperationsMemory;
-    use bridge_canister::runtime::state::config::ConfigStorage;
-    use bridge_canister::runtime::state::{SharedConfig, State};
-    use candid::Encode;
-    use ic_exports::ic_cdk::api::management_canister::bitcoin::Outpoint;
-    use ic_exports::ic_kit::MockContext;
-    use ic_stable_structures::MemoryId;
-    use snapbox::{assert_data_eq, str};
-
-    use super::*;
-    use crate::core::rune_inputs::mock::TestRuneInputProvider;
-    use crate::core::rune_inputs::{GetInputsError, RuneInput};
-    use crate::core::utxo_handler::test::TestUtxoHandler;
-    use crate::core::utxo_handler::UtxoHandlerError;
-    use crate::rune_info::RuneInfo;
-
-    fn op_memory() -> OperationsMemory<StableMemory> {
-        OperationsMemory {
-            id_counter: memory_by_id(MemoryId::new(1)),
-            incomplete_operations: memory_by_id(MemoryId::new(2)),
-            operations_log: memory_by_id(MemoryId::new(3)),
-            operations_map: memory_by_id(MemoryId::new(4)),
-            memo_operations_map: memory_by_id(MemoryId::new(5)),
-        }
-    }
-
-    fn config() -> SharedConfig {
-        Rc::new(RefCell::new(ConfigStorage::default(memory_by_id(
-            MemoryId::new(5),
-        ))))
-    }
-
-    fn test_state() -> RuntimeState<RuneBridgeOp> {
-        Rc::new(RefCell::new(State::default(op_memory(), config())))
-    }
-
-    fn sender() -> H160 {
-        H160::from_slice(&[1; 20])
-    }
-
-    fn rune_name(name: &str) -> RuneName {
-        RuneName::from_str(name).unwrap()
-    }
-
-    fn token_address(v: u8) -> H160 {
-        H160::from_slice(&[v; 20])
-    }
-
-    fn dst_tokens() -> HashMap<RuneName, H160> {
-        [
-            (rune_name("AAA"), token_address(2)),
-            (rune_name("A"), token_address(3)),
-            (rune_name("B"), token_address(4)),
-        ]
-        .into()
-    }
-
-    #[tokio::test]
-    async fn invalid_notification_type_is_noop() {
-        let notification = RuneDepositRequestData {
-            dst_address: sender(),
-            dst_tokens: dst_tokens(),
-            amounts: None,
-        };
-
-        let event = NotifyMinterEventData {
-            notification_type: MinterNotificationType::RescheduleOperation,
-            tx_sender: sender(),
-            user_data: Encode!(&notification).unwrap(),
-            memo: vec![],
-        };
-
-        let result = RuneBridgeOp::on_minter_notification(test_state(), event.clone()).await;
-        assert!(result.is_none());
-
-        let event = NotifyMinterEventData {
-            notification_type: MinterNotificationType::Other,
-            ..event
-        };
-        let result = RuneBridgeOp::on_minter_notification(test_state(), event.clone()).await;
-        assert!(result.is_none());
-    }
-
-    #[tokio::test]
-    async fn invalid_notification_payload_is_noop() {
-        let notification = RuneDepositRequestData {
-            dst_address: sender(),
-            dst_tokens: dst_tokens(),
-            amounts: None,
-        };
-        let mut data = Encode!(&notification).unwrap();
-        data.push(0);
-
-        let event = NotifyMinterEventData {
-            notification_type: MinterNotificationType::DepositRequest,
-            tx_sender: sender(),
-            user_data: data,
-            memo: vec![],
-        };
-
-        let result = RuneBridgeOp::on_minter_notification(test_state(), event.clone()).await;
-        assert!(result.is_none());
-
-        let event = NotifyMinterEventData {
-            user_data: vec![],
-            ..event
-        };
-        let result = RuneBridgeOp::on_minter_notification(test_state(), event.clone()).await;
-        assert!(result.is_none());
-    }
-
-    #[tokio::test]
-    async fn deposit_request_creates_correct_operation() {
-        let notification = RuneDepositRequestData {
-            dst_address: sender(),
-            dst_tokens: dst_tokens(),
-            amounts: None,
-        };
-        let data = Encode!(&notification).unwrap();
-
-        let event = NotifyMinterEventData {
-            notification_type: MinterNotificationType::DepositRequest,
-            tx_sender: sender(),
-            user_data: data,
-            memo: vec![],
-        };
-
-        let result = RuneBridgeOp::on_minter_notification(test_state(), event.clone()).await;
-        assert_eq!(
-            result,
-            Some(OperationAction::Create(
-                RuneBridgeOp::AwaitInputs {
-                    dst_address: sender(),
-                    dst_tokens: dst_tokens(),
-                    requested_amounts: None,
-                },
-                None
-            ))
-        )
-    }
-
-    #[tokio::test]
-    async fn deposit_request_adds_amounts_to_operation() {
-        let amounts: HashMap<RuneName, u128> = [(rune_name("AAA"), 1000)].into();
-        let notification = RuneDepositRequestData {
-            dst_address: sender(),
-            dst_tokens: dst_tokens(),
-            amounts: Some(amounts.clone()),
-        };
-        let data = Encode!(&notification).unwrap();
-
-        let event = NotifyMinterEventData {
-            notification_type: MinterNotificationType::DepositRequest,
-            tx_sender: sender(),
-            user_data: data,
-            memo: vec![],
-        };
-
-        let result = RuneBridgeOp::on_minter_notification(test_state(), event.clone()).await;
-        assert_eq!(
-            result,
-            Some(OperationAction::Create(
-                RuneBridgeOp::AwaitInputs {
-                    dst_address: sender(),
-                    dst_tokens: dst_tokens(),
-                    requested_amounts: Some(amounts),
-                },
-                None
-            ))
-        )
-    }
-
-    #[tokio::test]
-    async fn await_inputs_returns_error_if_no_inputs() {
-        let provider = TestRuneInputProvider::empty();
-        let result =
-            RuneBridgeOp::await_inputs(test_state(), &provider, sender(), dst_tokens(), None).await;
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Invalid result: {result:?}");
-        };
-
-        assert_data_eq!(message, str!["no inputs"])
-    }
-
-    #[tokio::test]
-    async fn await_inputs_returns_error_if_provider_returns_btc_error() {
-        let provider =
-            TestRuneInputProvider::err(GetInputsError::BtcAdapter("not available".to_string()));
-        let result =
-            RuneBridgeOp::await_inputs(test_state(), &provider, sender(), dst_tokens(), None).await;
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Invalid result: {result:?}");
-        };
-
-        assert_data_eq!(
-            message,
-            str![
-                "failed to get deposit inputs: failed to connect to IC BTC adapter: not available"
-            ]
-        )
-    }
-
-    #[tokio::test]
-    async fn await_inputs_returns_error_if_provider_returns_indexer_error() {
-        let provider = TestRuneInputProvider::err(GetInputsError::InsufficientConsensus {
-            received_responses: 0,
-            required_responses: 1,
-            checked_indexers: 0,
-        });
-        let result =
-            RuneBridgeOp::await_inputs(test_state(), &provider, sender(), dst_tokens(), None).await;
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Invalid result: {result:?}");
-        };
-
-        assert_data_eq!(
-            message,
-            str![
-                "failed to get deposit inputs: insufficient consensus from rune indexers: 0/1 responses received, 0 indexers checked"
-            ]
-        )
-    }
-
-    #[tokio::test]
-    async fn await_inputs_returns_error_if_provider_returns_consensus_error() {
-        let provider = TestRuneInputProvider::err(GetInputsError::IndexersDisagree {
-            indexer_responses: vec![("indexer_name".to_string(), "indexer_response".to_string())],
-        });
-        let result =
-            RuneBridgeOp::await_inputs(test_state(), &provider, sender(), dst_tokens(), None).await;
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Invalid result: {result:?}");
-        };
-
-        assert_data_eq!(
-            message,
-            str![[
-                r#"failed to get deposit inputs: rune indexers returned different result for same request: [("indexer_name", "indexer_response")]"#
-            ]]
-        )
-    }
-
-    fn rune_input(rune_name: &str, amount: u128) -> RuneInput {
-        RuneInput {
-            utxo: Utxo {
-                outpoint: Outpoint {
-                    txid: vec![],
-                    vout: 0,
-                },
-                value: 10_000,
-                height: 0,
-            },
-            runes: [(RuneName::from_str(rune_name).unwrap(), amount)].into(),
-        }
-    }
-
-    #[tokio::test]
-    async fn await_inputs_returns_error_if_wrong_amounts_one_utxo() {
-        let input = rune_input("A", 1000);
-        let provider = TestRuneInputProvider::with_input(input.clone());
-        let result = RuneBridgeOp::await_inputs(
-            test_state(),
-            &provider,
-            sender(),
-            dst_tokens(),
-            Some([(RuneName::from_str("B").unwrap(), 1000)].into()),
-        )
-        .await;
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Invalid result: {result:?}");
-        };
-
-        assert_data_eq!(message, str!["requested amounts {RuneName(Rune(1)): 1000} are not equal actual amounts {RuneName(Rune(0)): 1000}"]);
-
-        let input = rune_input("A", 1000);
-        let provider = TestRuneInputProvider::with_input(input.clone());
-        let result = RuneBridgeOp::await_inputs(
-            test_state(),
-            &provider,
-            sender(),
-            dst_tokens(),
-            Some([(RuneName::from_str("A").unwrap(), 2000)].into()),
-        )
-        .await;
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Invalid result: {result:?}");
-        };
-
-        assert_data_eq!(message, str!["requested amounts {RuneName(Rune(0)): 2000} are not equal actual amounts {RuneName(Rune(0)): 1000}"])
-    }
-
-    #[tokio::test]
-    async fn await_inputs_returns_error_if_wrong_amounts_multiple_utxos() {
-        let inputs = [rune_input("A", 1000), rune_input("B", 2000)];
-        let provider = TestRuneInputProvider::with_inputs(&inputs);
-        let result = RuneBridgeOp::await_inputs(
-            test_state(),
-            &provider,
-            sender(),
-            dst_tokens(),
-            Some([(RuneName::from_str("A").unwrap(), 1000)].into()),
-        )
-        .await;
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Invalid result: {result:?}");
-        };
-
-        assert_data_eq!(
-            message,
-            str!["requested amounts {RuneName(Rune(0)): 1000} are not equal actual amounts [..]"]
-        );
-    }
-
-    #[tokio::test]
-    async fn await_inputs_returns_error_if_no_token_address() {
-        let inputs = [rune_input("A", 1000)];
-        let provider = TestRuneInputProvider::with_inputs(&inputs);
-        let result = RuneBridgeOp::await_inputs(
-            test_state(),
-            &provider,
-            sender(),
-            [(RuneName::from_str("C").unwrap(), token_address(5))].into(),
-            None,
-        )
-        .await;
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Invalid result: {result:?}");
-        };
-
-        assert_data_eq!(message, str!["wrapped token address for rune A not found"]);
-    }
-
-    #[tokio::test]
-    async fn await_inputs_returns_correct_operation_single_input() {
-        let input = rune_input("A", 1000);
-        let provider = TestRuneInputProvider::with_input(input.clone());
-        let result =
-            RuneBridgeOp::await_inputs(test_state(), &provider, sender(), dst_tokens(), None).await;
-        assert_eq!(
-            result,
-            Ok(RuneBridgeOp::AwaitConfirmations {
-                dst_address: sender(),
-                utxo: input.utxo,
-                runes_to_wrap: vec![RuneToWrap {
-                    rune_info: provider.rune_info(&RuneName::from_str("A").unwrap()),
-                    amount: 1000,
-                    wrapped_address: token_address(3),
-                }],
-            })
-        );
-    }
-
-    #[tokio::test]
-    async fn await_inputs_returns_correct_operation_multiple_inputs() {
-        MockContext::new().inject();
-
-        let inputs = vec![rune_input("A", 1000), rune_input("B", 2000)];
-        let provider = TestRuneInputProvider::with_inputs(&inputs);
-        let state = test_state();
-        let result =
-            RuneBridgeOp::await_inputs(state.clone(), &provider, sender(), dst_tokens(), None)
-                .await;
-
-        let Ok(RuneBridgeOp::OperationSplit {
-            wallet_address,
-            new_operation_ids,
-        }) = result
-        else {
-            panic!("Incorrect operation returned")
-        };
-
-        for operation in new_operation_ids {
-            assert!(state.borrow().operations.get(operation).is_some());
-        }
-
-        assert_eq!(wallet_address, sender());
-    }
-
-    fn get_utxo() -> Utxo {
-        Utxo {
-            outpoint: Outpoint {
-                txid: vec![],
-                vout: 0,
-            },
-            value: 0,
-            height: 0,
-        }
-    }
-
-    fn get_to_wrap(count: usize) -> Vec<RuneToWrap> {
-        let mut result = vec![];
-        for _ in 0..count {
-            result.push(RuneToWrap {
-                rune_info: RuneInfo {
-                    name: RuneName::from_str("A").unwrap(),
-                    decimals: 0,
-                    block: 0,
-                    tx: 0,
-                },
-                amount: 0,
-                wrapped_address: H160::from_slice(&[1; 20]),
-            })
-        }
-
-        result
-    }
-
-    #[tokio::test]
-    async fn await_confirmations_utxo_not_found() {
-        let utxo_handler = TestUtxoHandler::with_error(UtxoHandlerError::UtxoNotFound);
-        let result = RuneBridgeOp::await_confirmations(
-            test_state(),
-            &utxo_handler,
-            sender(),
-            get_utxo(),
-            get_to_wrap(1),
-        )
-        .await;
-
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Wrong result: {result:?}");
-        };
-
-        assert_data_eq!(message, str!["requested utxo is not in the main branch"]);
-    }
-
-    #[tokio::test]
-    async fn await_confirmations_not_confirmed() {
-        let utxo_handler = TestUtxoHandler::with_error(UtxoHandlerError::NotConfirmed {
-            required_confirmations: 12,
-            current_confirmations: 5,
-        });
-        let result = RuneBridgeOp::await_confirmations(
-            test_state(),
-            &utxo_handler,
-            sender(),
-            get_utxo(),
-            get_to_wrap(1),
-        )
-        .await;
-
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Wrong result: {result:?}");
-        };
-
-        assert_data_eq!(
-            message,
-            str!["utxo is not confirmed, required 12, currently 5 confirmations"]
-        );
-    }
-
-    #[tokio::test]
-    async fn await_confirmations_btc_adapter_not_available() {
-        let utxo_handler =
-            TestUtxoHandler::with_error(UtxoHandlerError::BtcAdapter("btc error".to_string()));
-        let result = RuneBridgeOp::await_confirmations(
-            test_state(),
-            &utxo_handler,
-            sender(),
-            get_utxo(),
-            get_to_wrap(1),
-        )
-        .await;
-
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Wrong result: {result:?}");
-        };
-
-        assert_data_eq!(
-            message,
-            str!["failed to connect to IC BTC adapter: btc error"]
-        );
-    }
-
-    #[tokio::test]
-    async fn await_confirmations_utxo_already_used() {
-        let utxo_handler = TestUtxoHandler::already_used_utxo();
-        let result = RuneBridgeOp::await_confirmations(
-            test_state(),
-            &utxo_handler,
-            sender(),
-            get_utxo(),
-            get_to_wrap(1),
-        )
-        .await;
-
-        let Err(Error::FailedToProgress(message)) = result else {
-            panic!("Wrong result: {result:?}");
-        };
-
-        assert_data_eq!(message, str!["utxo is already used to create mint orders"]);
-    }
-
-    #[tokio::test]
-    async fn await_confirmations_one_mint_order() {
-        let utxo_handler = TestUtxoHandler::ok();
-        let result = RuneBridgeOp::await_confirmations(
-            test_state(),
-            &utxo_handler,
-            sender(),
-            get_utxo(),
-            get_to_wrap(1),
-        )
-        .await;
-
-        let Ok(operation) = result else {
-            panic!("Wrong result: {result:?}");
-        };
-
-        assert!(matches!(operation, RuneBridgeOp::SignMintOrder { .. }));
-    }
-
-    #[tokio::test]
-    async fn await_confirmations_multiple_mint_orders() {
-        MockContext::new().inject();
-
-        const COUNT: usize = 3;
-        let utxo_handler = TestUtxoHandler::ok();
-        let state = test_state();
-        let result = RuneBridgeOp::await_confirmations(
-            state.clone(),
-            &utxo_handler,
-            sender(),
-            get_utxo(),
-            get_to_wrap(COUNT),
-        )
-        .await;
-
-        let Ok(RuneBridgeOp::OperationSplit {
-            new_operation_ids,
-            wallet_address,
-        }) = result
-        else {
-            panic!("Wrong result: {result:?}");
-        };
-
-        assert_eq!(new_operation_ids.len(), COUNT);
-
-        for operation_id in new_operation_ids {
-            let operation = state.borrow().operations.get(operation_id).unwrap();
-            assert!(matches!(operation, RuneBridgeOp::SignMintOrder { .. }));
-        }
-
-        assert_eq!(wallet_address, sender());
-    }
-}
+mod tests;
